+++ kernel_tuner/core.py
@@ -291,7 +291,7 @@ class DeviceInterface(object):
             if any([skip_str in str(e) for skip_str in skippable_exceptions]):
                 logging.debug('benchmark fails due to runtime failure too many resources required')
                 if verbose:
-                    print("skipping config", instance.name, "reason: too many resources requested for launch")
+                    print(f"skipping config {util.get_instance_string(instance.params)} reason: too many resources requested for launch")
             else:
                 logging.debug('benchmark encountered runtime failure: ' + str(e))
                 print("Error while benchmarking:", instance.name)
@@ -408,7 +408,7 @@ class DeviceInterface(object):
             if any(msg in str(e) for msg in shared_mem_error_messages):
                 logging.debug('compile_kernel failed due to kernel using too much shared memory')
                 if verbose:
-                    print("skipping config", instance.name, "reason: too much shared memory used")
+                    print(f"skipping config {util.get_instance_string(instance.params)} reason: too much shared memory used")
             else:
                 logging.debug('compile_kernel failed due to error: ' + str(e))
                 print("Error while compiling:", instance.name)
@@ -438,7 +438,6 @@ class DeviceInterface(object):
 
     def create_kernel_instance(self, kernel_source, kernel_options, params, verbose):
         """create kernel instance from kernel source, parameters, problem size, grid divisors, and so on"""
-        instance_string = util.get_instance_string(params)
         grid_div = (kernel_options.grid_div_x, kernel_options.grid_div_y, kernel_options.grid_div_z)
 
         #insert default block_size_names if needed
@@ -449,7 +448,7 @@ class DeviceInterface(object):
         threads, grid = util.setup_block_and_grid(kernel_options.problem_size, grid_div, params, kernel_options.block_size_names)
         if np.prod(threads) > self.dev.max_threads:
             if verbose:
-                print("skipping config", instance_string, "reason: too many threads per block")
+                print(f"skipping config {util.get_instance_string(params)} reason: too many threads per block")
             return None
 
         #obtain the kernel_string and prepare additional files, if any
@@ -508,29 +507,29 @@ def _default_verify_function(instance, answer, result_host, atol, verbose):
         if answer[i] is not None:    #skip None elements in the answer list
             if isinstance(answer[i], (np.ndarray, cp.ndarray)) and isinstance(arg, (np.ndarray, cp.ndarray)):
                 if answer[i].dtype != arg.dtype:
-                    raise TypeError("Element " + str(i) + " of the expected results list is not of the same dtype as the kernel output: " +
+                    raise TypeError(f"Element {i} of the expected results list is not of the same dtype as the kernel output: " +
                                     str(answer[i].dtype) + " != " + str(arg.dtype) + ".")
                 if answer[i].size != arg.size:
-                    raise TypeError("Element " + str(i) + " of the expected results list has a size different from " + "the kernel argument: " +
+                    raise TypeError(f"Element {i} of the expected results list has a size different from " + "the kernel argument: " +
                                     str(answer[i].size) + " != " + str(arg.size) + ".")
             elif isinstance(answer[i], torch.Tensor) and isinstance(arg, torch.Tensor):
                 if answer[i].dtype != arg.dtype:
-                    raise TypeError("Element " + str(i) + " of the expected results list is not of the same dtype as the kernel output: " +
+                    raise TypeError(f"Element {i} of the expected results list is not of the same dtype as the kernel output: " +
                                     str(answer[i].dtype) + " != " + str(arg.dtype) + ".")
                 if answer[i].size() != arg.size():
-                    raise TypeError("Element " + str(i) + " of the expected results list has a size different from " + "the kernel argument: " +
+                    raise TypeError(f"Element {i} of the expected results list has a size different from " + "the kernel argument: " +
                                     str(answer[i].size) + " != " + str(arg.size) + ".")
 
             elif isinstance(answer[i], np.number) and isinstance(arg, np.number):
                 if answer[i].dtype != arg.dtype:
-                    raise TypeError("Element " + str(i) + " of the expected results list is not the same as the kernel output: " + str(answer[i].dtype) +
+                    raise TypeError(f"Element {i} of the expected results list is not the same as the kernel output: " + str(answer[i].dtype) +
                                     " != " + str(arg.dtype) + ".")
             else:
                 #either answer[i] and argument have different types or answer[i] is not a numpy type
                 if not isinstance(answer[i], (np.ndarray, cp.ndarray, torch.Tensor)) or not isinstance(answer[i], np.number):
-                    raise TypeError("Element " + str(i) + " of expected results list is not a numpy/cupy ndarray, torch Tensor or numpy scalar.")
+                    raise TypeError(f"Element {i} of expected results list is not a numpy/cupy ndarray, torch Tensor or numpy scalar.")
                 else:
-                    raise TypeError("Element " + str(i) + " of expected results list and kernel arguments have different types.")
+                    raise TypeError(f"Element {i} of expected results list and kernel arguments have different types.")
 
     def _ravel(a):
         if hasattr(a, 'ravel') and len(a.shape) > 1: