+++ CHANGELOG.md
@@ -5,13 +5,18 @@ This project adheres to [Semantic Versioning](http://semver.org/).
 ## Unreleased
 
 ### Added
+- new optimization strategies: dual annealing, greedly ILS, ordered greedy MLS, greedy MLS
+
+## [0.4.1] - 2021-09-10
+### Added
 - support for PyTorch Tensors as input data type for kernels
 - support for smem_args in run_kernel
 - support for (lambda) function and string for dynamic shared memory size
+- a new Bayesian Optimization strategy
 
 ### Changed
 - optionally store the kernel_string with store_results
-
+- improved reporting of skipped configurations
 
 ## [0.4.0] - 2021-04-09
 ### Added